Refactor TransferAccordion tests
Jira ticket: CAMS-338
